Generate sample tests based on framework
